+++ src/Monoid.ts
@@ -84,9 +84,9 @@ export const monoidVoid: Monoid<void> = {
 /**
  * @since 2.0.0
  */
-export function fold<A>(M: Monoid<A>): (as: Array<A>) => A {
-  const foldSemigroupM = foldSemigroup(M)
-  return as => foldSemigroupM(M.empty, as)
+export function fold<A>(M: Monoid<A>): (as: ReadonlyArray<A>) => A {
+  const foldM = foldSemigroup(M)
+  return as => foldM(M.empty, as)
 }
 
 /**