+++ src/Semigroup.ts
@@ -19,7 +19,7 @@ export interface Semigroup<A> extends Magma<A> {}
 /**
  * @since 2.0.0
  */
-export function fold<A>(S: Semigroup<A>): (a: A, as: Array<A>) => A {
+export function fold<A>(S: Semigroup<A>): (a: A, as: ReadonlyArray<A>) => A {
   return (a, as) => as.reduce(S.concat, a)
 }